refactor(bytecode-utils): isolate EraVM/zksolc auxdata into eravm.ts
Split the interleaved zksolc logic out of the shared bytecode.ts into a
clean 3-file DAG:

- cbor.ts:  generic CBOR-tail primitives + AuxdataStyle enum (no EraVM
            knowledge). Holds splitCborAuxdata and the extracted
            decodeSolidityCborObject that EraVM reuses.
- eravm.ts: all EraVM-specific logic — splitEraVmAuxdata, the new
            decodeEraVmAuxdata (padding strip), and ERA_VM_* constants.
            Depends only on cbor.ts.
- bytecode.ts: public facade — decode/splitAuxdata dispatch and the
            Solidity/Vyper branches. Its only EraVM awareness is two thin
            dispatch lines; it re-exports the CBOR public API so the
            package's exported surface is unchanged.

Move the four EraVM test cases + fixtures into a dedicated eravm.spec.ts.
Update the stale bytecode.ts reference in ZKSOLC.md to eravm.ts.

No behavior change: all 22 bytecode-utils tests pass and lib-sourcify
consumers need no changes.
